+++ src/HttpException.php
@@ -20,10 +20,10 @@ class HttpException extends Exception
      *
      * @param string $message [optional] The Exception message to throw.
      * @param int $code [optional] The Exception code.
-     * @param Exception $previous [optional] The previous throwable used for the exception chaining.
+     * @param \Throwable|\Exception $previous [optional] The previous throwable used for the exception chaining.
      * @param array $parameters [optional] Array of parameters returned from API.
      */
-    public function __construct($message = '', $code = 0, Exception $previous = null, $parameters = [])
+    public function __construct($message = '', $code = 0, $previous = null, $parameters = [])
     {
         $this->parameters = $parameters;