+++ tests/BotApiTest.php
@@ -4,6 +4,7 @@ namespace TelegramBot\Api\Test;
 
 use PHPUnit\Framework\TestCase;
 use TelegramBot\Api\BotApi;
+use TelegramBot\Api\Http\HttpClientInterface;
 use TelegramBot\Api\Types\ArrayOfUpdates;
 use TelegramBot\Api\Types\Update;
 
@@ -108,16 +109,18 @@ class BotApiTest extends TestCase
      */
     public function testGetUpdates($updates)
     {
-        $mock = $this->getMockBuilder(BotApi::class)
-            ->setMethods(['call'])
-            ->enableOriginalConstructor()
-            ->setConstructorArgs(['testToken'])
-            ->getMock();
+        $httpClient = $this->createHttpClient();
+        $botApi = $this->createBotApi($httpClient);
 
-        $mock->expects($this->once())->method('call')->willReturn($updates);
+        $httpClient
+            ->expects($this->once())
+            ->method('request')
+            ->willReturn($updates)
+        ;
+
+        $result = $botApi->getUpdates();
 
         $expectedResult = ArrayOfUpdates::fromResponse($updates);
-        $result = $mock->getUpdates();
 
         $this->assertEquals($expectedResult, $result);
 
@@ -125,5 +128,21 @@ class BotApiTest extends TestCase
             $this->assertInstanceOf(Update::class, $item);
             $this->assertEquals($expectedResult[$key], $item);
         }
+    }
+
+    /**
+     * @return HttpClientInterface&\PHPUnit\Framework\MockObject\MockObject
+     */
+    private function createHttpClient()
+    {
+        /** @var HttpClientInterface $httpClient */
+        $httpClient = $this->createMock(HttpClientInterface::class);
+
+        return $httpClient;
+    }
+
+    private function createBotApi(HttpClientInterface $httpClient)
+    {
+        return new BotApi('token', null, $httpClient);
     }
 }