+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All Notable changes to `PHP Telegram Bot Api` will be documented in this file
 
-## 2.6.0 - YYYY-MM-DD
+## 3.0.0 - YYYY-MM-DD
 - Add `\TelegramBot\Api\Types\Update::$myChatMember` field
 - Add `\TelegramBot\Api\Types\Update::$chatMember` field
 - Add `\TelegramBot\Api\Types\Update::$chatJoinRequest` field
@@ -15,6 +15,7 @@ All Notable changes to `PHP Telegram Bot Api` will be documented in this file
 - Add support for local bot API server
 - Add method `\TelegramBot\Api\BotApi::validateWebAppData` to validate `window.Telegram.WebApp.initData`
 - Add `\TelegramBot\Api\Types\Message::$videoNote` field
+- Drop php < 8.1
 
 ## 2.5.0 - 2023-08-09